+++ internal/provider/provider.go
@@ -78,6 +78,7 @@ func (p *Provider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewRepositoryActionsSecretResource,
 		NewRepositoryActionsVariableResource,
+		NewRepositoryPushMirrorResource,
 	}
 }