+++ CHANGELOG.md
@@ -11,5 +11,6 @@ All notable changes to this project will be documented in this file.
 - Added repositories data-source.
 - Added repository actions secret resource.
 - Added repository actions variable resource.
+- Added repository push mirror resource.
 - Added teams data-source.
 - Added users data-source.